Provides backend support to DWIN_SET v0.5.0
Converts the list function on the Print Menu to look and behave the same way as the Custom Macro screens.

+++ klippy/extras/t5uid1/t5uid1.py
@@ -689,7 +689,7 @@ class T5UID1:
         self._scroll_index = index
         try: # Find the file path in _files based on the index + _scroll_index 
             file_path = self._files[self._scroll_index] 
-            if file_path is not None:
+            if file_path is not None and file_path != "None":
                 # Delete the file
                 os.remove(file_path)
                 logging.info(f"Deleted file: {file_path}")
